Generate a new cohort matched cohort from a preexisting target cohort. The new cohort will contain individuals not included in the target cohort with same year of birth (matchYearOfBirth = TRUE) and same sex (matchSex = TRUE).

Generate a new cohort matched cohort from a preexisting target cohort. The new cohort will contain individuals not included in the target cohort with same year of birth (matchYearOfBirth = TRUE) and same sex (matchSex = TRUE).

Usage

generateMatchedCohortSet(
  cdm,
  name,
  targetCohortName,
  targetCohortId = NULL,
  matchSex = TRUE,
  matchYearOfBirth = TRUE,
  ratio = 1
)

Arguments

cdm

A cdm reference object.

name

Name of the new generated cohort set.

targetCohortName

Name of the target cohort to match.

targetCohortId

Cohort definition id to match from the target cohort. If NULL all the cohort definition id present in the target cohort will be matched.

matchSex

Whether to match in sex.

matchYearOfBirth

Whether to match in year of birth.

ratio

Number of allowed matches per individual in the target cohort.

Value

A cdm reference object that contains the new generated cohort set.

Examples

library(DrugUtilisation)
library(CohortConstructor)
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
cdm <- mockDrugUtilisation(numberIndividuals = 100)
cdm <- cdm %>%
  generateMatchedCohortSet(name = "new_matched_cohort",
                           targetCohortName = "cohort1",
                           targetCohortId = c(1,2),
                           matchSex = TRUE,
                           matchYearOfBirth = TRUE,
                           ratio = 2)
cdm$new_matched_cohort
#> # Source:   table<new_matched_cohort> [10 x 4]
#> # Database: DuckDB v0.9.2 [unknown@Linux 6.2.0-1018-azure:R 4.3.2/:memory:]
#>    cohort_definition_id subject_id cohort_start_date cohort_end_date
#>                   <int>      <int> <date>            <date>         
#>  1                    1         87 2019-08-23        2020-01-05     
#>  2                    2         83 2019-08-23        2022-06-22     
#>  3                    1         79 2017-08-24        2019-02-12     
#>  4                    2         55 2017-08-24        2021-02-20     
#>  5                    3         55 2018-02-10        2018-08-03     
#>  6                    4         79 2018-02-10        2022-04-04     
#>  7                    3         86 2014-10-28        2014-11-19     
#>  8                    4         42 2014-10-28        2021-03-29     
#>  9                    1         96 2010-08-26        2015-04-13     
#> 10                    2         78 2010-08-26        2012-02-25