Generate a combination cohort set between the intersection of different cohorts.

intersectCohorts() combines different cohort entries, with those records that overlap combined and kept. Cohort entries are when an individual was in both of the cohorts.

Usage

intersectCohorts(
  cohort,
  cohortId = NULL,
  gap = 0,
  mutuallyExclusive = FALSE,
  returnOnlyComb = FALSE,
  name = tableName(cohort)
)

Arguments

cohort

A cohort table in a cdm reference.

cohortId

IDs of the cohorts to include. If NULL all cohorts will be considered. Cohorts not included will be removed from the cohort set.

gap

Number of days between two subsequent cohort entries to be merged in a single cohort record.

mutuallyExclusive

Whether the generated cohorts are mutually exclusive or not.

returnOnlyComb

Whether to only get the combination cohort back

name

Name of the new cohort with the demographic requirements.

Value

A cohort table.

Examples

# \donttest{
library(CohortConstructor)

cdm <- mockCohortConstructor(nPerson = 100)

cdm$cohort3 <- intersectCohorts(
  cohort = cdm$cohort2,
  name = "cohort3",
)

settings(cdm$cohort3)
#> # A tibble: 3 × 6
#>   cohort_definition_id cohort_name    cohort_1 cohort_2 mutually_exclusive   gap
#>                  <int> <chr>             <dbl>    <dbl> <lgl>              <dbl>
#> 1                    1 cohort_1              1        0 FALSE                  0
#> 2                    2 cohort_2              0        1 FALSE                  0
#> 3                    3 cohort_1_coho…        1        1 FALSE                  0

# }