Generate a new cohort matched cohort

matchCohorts() generate a new cohort matched to individuals in an existing cohort. Individuals can be matched based on year of birth and sex.

Usage

matchCohorts(
  cohort,
  cohortId = NULL,
  matchSex = TRUE,
  matchYearOfBirth = TRUE,
  ratio = 1,
  name = tableName(cohort)
)

Arguments

cohort

A cohort table in a cdm reference.

cohortId

IDs of the cohorts to include. If NULL all cohorts will be considered. Cohorts not included will be removed from the cohort set.

matchSex

Whether to match in sex.

matchYearOfBirth

Whether to match in year of birth.

ratio

Number of allowed matches per individual in the target cohort.

name

Name of the new generated cohort set.

Value

A cohort table.

Examples

# \donttest{
library(CohortConstructor)
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
cdm <- mockCohortConstructor(nPerson = 200)
cdm$new_matched_cohort <- cdm$cohort2 |>
  matchCohorts(
    name = "new_matched_cohort",
    cohortId = 2,
    matchSex = TRUE,
    matchYearOfBirth = TRUE,
    ratio = 1)
#> Starting matching
#> Warning: Multiple records per person detected. The matchCohorts() function is designed
#> to operate under the assumption that there is only one record per person within
#> each cohort. If this assumption is not met, each record will be treated
#> independently. As a result, the same individual may be matched multiple times,
#> leading to inconsistent and potentially misleading results.
#> ℹ Creating copy of target cohort.
#> • 1 cohort to be matched.
#> ℹ Creating controls cohorts.
#> ℹ Excluding cases from controls
#> • Matching by gender_concept_id and year_of_birth
#> • Removing controls that were not in observation at index date
#> • Excluding target records whose pair is not in observation
#> • Adjusting ratio
#> Binding both cohorts
#> ✔ Done
cdm$new_matched_cohort
#> # Source:   table<main.new_matched_cohort> [?? x 5]
#> # Database: DuckDB v1.0.0 [unknown@Linux 6.5.0-1024-azure:R 4.4.1/:memory:]
#>    cohort_definition_id subject_id cohort_start_date cohort_end_date cluster_id
#>                   <int>      <int> <date>            <date>               <dbl>
#>  1                    1        118 1975-11-28        1981-05-30               6
#>  2                    1         56 2007-09-28        2008-07-15              18
#>  3                    1         20 1990-06-07        1993-11-05              27
#>  4                    1         84 1999-04-06        2000-04-12              42
#>  5                    1         94 2017-11-01        2018-07-05              59
#>  6                    1        120 2009-12-15        2010-05-07              80
#>  7                    1         62 2017-04-27        2018-07-04              73
#>  8                    1         83 2011-02-07        2011-02-26              93
#>  9                    1        105 2001-10-24        2002-04-20             106
#> 10                    2        133 2007-09-28        2009-09-22              18
#> # ℹ more rows
# }