Restrict cohort on age

Restrict cohort on age

Usage

requireAge(cohort, indexDate = "cohort_start_date", ageRange = list(c(0, 150)))

Arguments

cohort

A cohort table in a cdm reference

indexDate

Variable in cohort that contains the date to compute the demographics characteristics on which to restrict on.

ageRange

A list of minimum and maximum age

Value

The cohort table with only records for individuals satisfying the age requirement

Examples

library(DrugUtilisation)
library(CohortConstructor)
cdm <- mockDrugUtilisation(numberIndividuals = 100)
cdm$cohort1 %>%
  requireAge(indexDate = "cohort_start_date",
             ageRange = list(c(18, 65)))
#> # Source:   SQL [?? x 4]
#> # Database: DuckDB v0.9.2 [unknown@Linux 6.2.0-1018-azure:R 4.3.2/:memory:]
#>    cohort_definition_id subject_id cohort_start_date cohort_end_date
#>                   <int>      <int> <date>            <date>         
#>  1                    1          2 2018-05-29        2018-06-10     
#>  2                    2          3 2001-08-11        2007-04-14     
#>  3                    2          6 2010-11-19        2011-05-04     
#>  4                    2          8 2020-09-29        2020-12-09     
#>  5                    1         11 2021-08-15        2021-09-26     
#>  6                    3         12 2020-09-20        2020-12-11     
#>  7                    3         13 1981-12-15        1982-10-31     
#>  8                    1         14 1976-03-20        1978-11-23     
#>  9                    1         15 2014-02-01        2014-12-31     
#> 10                    3         17 2006-07-30        2007-02-15     
#> # ℹ more rows