Restrict cohort on patient demographics

Restrict cohort on patient demographics

Usage

requireDemographics(
  cohort,
  indexDate = "cohort_start_date",
  ageRange = list(c(0, 150)),
  sex = c("Both"),
  minPriorObservation = 0,
  minFutureObservation = 0
)

Arguments

cohort

A cohort table in a cdm reference

indexDate

Variable in cohort that contains the date to compute the demographics characteristics on which to restrict on.

ageRange

A list of minimum and maximum age

sex

Can be "Both", "Male" or "Female". If one of the latter, only those with that sex will be included.

minPriorObservation

A minimum number of prior observation days in the database.

minFutureObservation

A minimum number of future observation days in the database.

Value

The cohort table with only records for individuals satisfying the demographic requirements

Examples

library(DrugUtilisation)
library(CohortConstructor)
cdm <- mockDrugUtilisation(numberIndividuals = 100)
cdm$cohort1 %>%
  requireDemographics(indexDate = "cohort_start_date",
                      ageRange = list(c(18, 65)),
                      sex = "Female",
                      minPriorObservation = 365)
#> # Source:   SQL [?? x 4]
#> # Database: DuckDB v0.9.2 [unknown@Linux 6.2.0-1018-azure:R 4.3.2/:memory:]
#>    cohort_definition_id subject_id cohort_start_date cohort_end_date
#>                   <int>      <int> <date>            <date>         
#>  1                    1          2 2018-05-29        2018-06-10     
#>  2                    3         17 2006-07-30        2007-02-15     
#>  3                    1         18 2003-03-14        2004-05-08     
#>  4                    2         27 2020-05-27        2020-06-26     
#>  5                    2         30 1988-05-24        1991-03-27     
#>  6                    1         32 2007-05-08        2007-11-29     
#>  7                    1         34 2011-10-26        2014-07-07     
#>  8                    1         48 1995-07-23        1997-12-11     
#>  9                    3         51 2021-02-15        2022-05-02     
#> 10                    1         53 1974-10-11        1978-03-06     
#> # ℹ more rows